+++ src/rgb/iter.rs
@@ -3,17 +3,17 @@ use crate::ColorTuple;
 
 pub struct RgbIter {
   ind: usize,
-  vals: [Option<f32>; 4],
+  vals: [Option<f64>; 4],
 }
 
 impl RgbIter {
-  pub fn new(t: ColorTuple, a: Option<f32>) -> RgbIter {
+  pub fn new(t: ColorTuple, a: Option<f64>) -> RgbIter {
     RgbIter { ind: 0, vals: [Some(t.0), Some(t.1), Some(t.2), a] }
   }
 }
 
 impl std::iter::Iterator for RgbIter {
-  type Item = f32;
+  type Item = f64;
   fn next(&mut self) -> Option<Self::Item> {
     match self.ind {
       0...3 => {
@@ -27,7 +27,7 @@ impl std::iter::Iterator for RgbIter {
 }
 
 impl<'a> std::iter::IntoIterator for &'a Rgb {
-  type Item = f32;
+  type Item = f64;
   type IntoIter = RgbIter;
   fn into_iter(self) -> RgbIter {
     self.iter()
@@ -35,7 +35,7 @@ impl<'a> std::iter::IntoIterator for &'a Rgb {
 }
 
 impl std::iter::IntoIterator for Rgb {
-  type Item = f32;
+  type Item = f64;
   type IntoIter = RgbIter;
   fn into_iter(self) -> RgbIter {
     self.iter()
